style: improve code & documentation
